requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
tzlocal<3
pytz
requests
python-dotenv